apocalypseinfrastructure/pipes/requirements.txt
pyyaml
pillow   # will be needed next